[FEATURE] Added tests to new __call method in property path
